+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Make single HTTP request when searching for profile/post by URL.
 - Reduced compilation time by disabling unused Actix features.
 - Don't panic if compatible ID is stored in database.
+- Validate conversation audience URI before saving it to database.
 
 ### Fixed